requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
aniso8601==9.0.1
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
bcrypt==3.2.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==0.9.3
colorama==0.4.6
cryptography==38.0.3
decorator==4.4.2
dnspython==2.2.1
email-validator==1.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
frozenlist==1.3.3
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy
packaging==21.3
pathspec==0.10.2
pg8000==1.29.3
Pillow==9.3.0
platformdirs==2.5.4
proglog==0.1.10
protobuf==4.21.9
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.0
py-topping
stripe 